update reforce func

+++ installer.sh
@@ -496,19 +496,23 @@ uninstall (){
   echo -e "$ORANGE""$BOLD""Consider running " "$CYAN""\$docker system prune""$NC"
 
   # delete/uninstall EMBA
-  if [ -f ./emba/install.log ]; then
-    rm ./emba/install.log
+  if [[ $REFORCE -eq 1 ]]; then
+    sudo -u "${SUDO_USER:-${USER}}" git submodule update
+  else
+    if [ -f ./emba/install.log ]; then
+      rm ./emba/install.log
+    fi
+    if [[ $(sudo -u "${SUDO_USER:-${USER}}" git submodule foreach git status --porcelain --untracked-files=no) ]]; then
+      echo -e "[!!]$RED""$BOLD""EMBA changes detected - please commit them...otherwise they will be lost""$NC"
+      read -p "If you know what you are doing you can press any key to continue ..." -n1 -s -r
+    fi
+    if [[ -d ./emba/external ]]; then
+      rm -r ./emba/external/
+    fi
+    sudo -u "${SUDO_USER:-${USER}}" git submodule foreach git reset --hard
+    sudo -u "${SUDO_USER:-${USER}}" git submodule deinit --all -f
   fi
-  if [[ $(sudo -u "${SUDO_USER:-${USER}}" git submodule foreach git status --porcelain --untracked-files=no) ]]; then
-    echo -e "[!!]$RED""$BOLD""EMBA changes detected - please commit them...otherwise they will be lost""$NC"
-    read -p "If you know what you are doing you can press any key to continue ..." -n1 -s -r
-  fi
-  if [[ -d ./emba/external ]]; then
-    rm -r ./emba/external/
-  fi
-  sudo -u "${SUDO_USER:-${USER}}" git submodule foreach git reset --hard
-  sudo -u "${SUDO_USER:-${USER}}" git submodule deinit --all -f
-
+  
   # stop&reset daemon
   if [[ "$WSL" -ne 1 ]]; then
     uninstall_daemon
